requirements.txt
requests
pandas
pyarrow
boto3
python-dotenv